Relationships__pwd, Protecting video routes

+++ controllers/videoController.js
@@ -1,6 +1,7 @@
 import routes from '../routes';
 import Video from '../models/Video';
 
+// Home
 export const home = async (req, res) => {
   try {
     const videos = await Video.find({});
@@ -11,59 +12,68 @@ export const home = async (req, res) => {
   }
 };
 
+// Search
 export const search = async (req, res) => {
   const {
     query: { term: searchingBy },
   } = req;
   let videos = [];
-  try{
-    videos = await Video.find({title: { $regex:searchingBy, $options:'i' }});
-  } catch(error){
+  try {
+    videos = await Video.find({
+      title: { $regex: searchingBy, $options: 'i' },
+    });
+  } catch (error) {
     console.log(error);
-  } 
-  res.render('search', { pageTitle: 'Search', searchingBy, videos});
-  
+  }
+  res.render('search', { pageTitle: 'Search', searchingBy, videos });
 };
 
+// Upload
 export const getUpload = (req, res) =>
   res.render('upload', { pageTitle: 'Upload' });
 
 export const postUpload = async (req, res) => {
   const {
     body: { title, description },
-    file: { path },
+    file: { location },
   } = req;
   const newVideo = await Video.create({
-    fileUrl: path,
+    fileUrl: location,
     title,
     description,
+    creator: req.user.id,
   });
-  console.log(newVideo);
-  // To Do: Upload and save video
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
+// Video Detail
 export const videoDetail = async (req, res) => {
   const {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate('creator');
     console.log(video);
     res.render('videoDetail', { pageTitle: video.title, video });
   } catch (error) {
-    console.log(error);
     res.redirect(routes.home);
   }
 };
 
+// Edit Video
 export const getEditVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -82,13 +92,19 @@ export const postEditVideo = async (req, res) => {
   }
 };
 
+// Delete Video
 export const deleteVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
 
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
